Fix compilation issue

+++ 0x1E-search_algorithms/102-interpolation.c
@@ -32,5 +32,5 @@ int interpolation_search(int *array, size_t size, int value)
 		else
 			high = pos - 1;
 	}
-	return (value == array[low] ? low : -1);
+	return (value == array[low] ? (int)low : -1);
 }
